backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# Security & Authentication
Werkzeug==3.0.1
PyJWT==2.8.0

# Environment & Configuration
python-dotenv==1.0.0

# Database (for future use)
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1

# Development Tools
gunicorn==21.2.0
requests==2.31.0

# Optional: Firebase integration
firebase-admin==6.4.0

# Optional: Input validation
marshmallow==3.20.1